ios: add xcrun shim for maestro cli to find simctl
